backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic-settings==2.2.1
